Add support for other in select and multi select
Add Other member to Select and MultiSelect structs.
If set to true another choice will appear, after all supplied options,
named Other, if selected the user is prompted to supply their input.
Used to have a set of choices but allowing the user to enter their own
value for the prompt.

This commit only includes tests for the output, not for any interaction.