allow to change the editor behavior
Add two fields `HideDefault` and `AppendDefaul` to `Editor`.

If `HideDefault` is `true`, the default value is not shown
before the launch of the editor. By default this value is `false`.

If `AppendDefault` is `true`, the default value is written to the temporary
file before the launch of the editor and `Prompt()` returns the empty
string if the contents of the temporary file is empty.
By default this value is `false`.