Add MultiInput prompt
This commit adds a MultiInput prompt allowing a user to enter multiple
distinct values for a question. These are then returned as a []string.
For example:

	? What are your friend's names?

Then the user is prompted for the first answer, then the second and so
on.

	? #1: Batman
	? #2: Joker
	? #3:

Then once finished they enter an empty line and the question will exit,
outputting the answers.

	? What are your friend's names? [Batman Joker]