Transparently handle errors from go-expect calls
Tests that used to invoke methods on a `*expect.Console` instance now
invoke same-named methods that have error handling built in so that the
associated test is properly aborted if any of the methods happen to
fail. This kind of error handling was easier to add than adding explicit
error handling to every `c.Expect*()` and `c.Send*()` call site.